Tests for DelayedAMPClient.
